INT-WTHR-8 - fixed for requested changes
