+++ README.md
@@ -1,1 +1,11 @@
-# weather_notifier
+# Implementation  weather push-notifications with darsky
+
+Service for continuous weather updates.
+
+Environment variables with default values:
+
+    LATITUDE       = 53.9045
+    LONGITUDE      = 27.5615   
+    SYNC_INTERVAL  = 5s
+
+    API_KEY and API_URL does not have a default value and must be provided.
